pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "keyboard>=0.13.5",
    "paramiko>=3.5.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
